Refactor: 불필요한 파일 삭제 및 Pagenation 폴더 정리
