--- src/apis/dataAccess.js
@@ -1,1 +1,0 @@
-// 임시 폴더입니다. 데이터 fetch 등 api 함수 등의 로직을 저장하는 폴더입니다.
